app/requirements.txt
django==2.2.12
djongo==1.3.2
invoke==0.10.0
djangorestframework==3.11.0
djangorestframework-xml==2.0.0
django-cors-headers==3.2.1
sqlparse==0.2.4
rest-meets-djongo==0.0.13
gunicorn==20.0.4